+++ src/ShutdownHandler/ShutdownHandler.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {{addTask: (task: () => Promise<void>) => Promise<void>, shutdown: (exit_code: number | null) => Promise<void>}} ShutdownHandler
+ */